refactor: reorder cipher permutation module
